Finish training and predict

+++ KMeans.h
@@ -2,17 +2,17 @@
 class KMeans
 {
 public:
-	KMeans(int,int,double);
+	KMeans(int,int,float);
 	~KMeans();
-	void Cluster(double*, int, int*);
-	double PredictLabel(double* v1, int* ppredict_label);
+	void Cluster(float*, int, int*);
+	float PredictLabel(float* v1, int* ppredict_label);
 	void PrintMeans();
-	double* GetMean(int index)	{ return m_means[index]; }
+	float* GetMean(int index)	{ return m_means[index]; }
 private:
 	int m_clusterNumber; 
 	int m_dimensionNumber; 
-	double m_endCondition;
-	double **m_means;
-	double Distance(const double*, const double*, int);
+	float m_endCondition;
+	float **m_means;
+	float Distance(const float*, const float*, int);
 
 };
