+++ libKMeans.cpp
@@ -3,16 +3,16 @@
 #include <stdio.h>
 #include <assert.h>  
 #include <Math.h>
-KMeans::KMeans(int dimensionNumber, int clusterNumber, double endCondition)
+KMeans::KMeans(int dimensionNumber, int clusterNumber, float endCondition)
 {
 	m_clusterNumber = clusterNumber;
 	m_dimensionNumber = dimensionNumber;
 	m_endCondition = endCondition;
-	m_means = new double*[m_clusterNumber];
+	m_means = new float*[m_clusterNumber];
 	for (int i = 0; i < m_clusterNumber; i++)
 	{
-		m_means[i] = new double[m_dimensionNumber];
-		memset(m_means[i], 0, sizeof(double) * m_dimensionNumber);
+		m_means[i] = new float[m_dimensionNumber];
+		memset(m_means[i], 0, sizeof(float) * m_dimensionNumber);
 	}
 }
 KMeans::~KMeans()
@@ -23,9 +23,9 @@ KMeans::~KMeans()
 	}
 	delete[] m_means;
 }
-double KMeans::Distance(const double* v1, const double* v2, int dimensionNumber)
+float KMeans::Distance(const float* v1, const float* v2, int dimensionNumber)
 {
-	double distance = 0;
+	float distance = 0;
 	for (int d = 0; d < dimensionNumber; d++)
 	{
 		distance += (v1[d] - v2[d]) * (v1[d] - v2[d]);
@@ -33,11 +33,11 @@ double KMeans::Distance(const double* v1, const double* v2, int dimensionNumber)
 	return sqrt(distance);
 }
 //return distance
-double KMeans::PredictLabel(double* v1, int* ppredict_label)
+float KMeans::PredictLabel(float* v1, int* ppredict_label)
 {
-	double distance = -1;
+	float distance = -1;
 	for (int i = 0; i < m_clusterNumber; i++){
-		double temp = Distance(v1, m_means[i], m_dimensionNumber);
+		float temp = Distance(v1, m_means[i], m_dimensionNumber);
 		if (temp < distance || distance == -1){
 			distance = temp;
 			*ppredict_label = i;
@@ -45,37 +45,37 @@ double KMeans::PredictLabel(double* v1, int* ppredict_label)
 	}
 	return distance;
 }
-void KMeans::Cluster(double *data, int size, int *Label)
+void KMeans::Cluster(float *data, int size, int *Label)
 {
 	assert(size >= m_clusterNumber);
 	//init m_means[] from data
-	double* temp_mean = new double[m_dimensionNumber];
+	float* temp_mean = new float[m_dimensionNumber];
 	for (int i = 0; i < m_clusterNumber; i++){
 		//select選出index為0, 1*(size / m_clusterNumber), 2*(size / m_clusterNumber)當作初始化群心
 		int select = i * size / m_clusterNumber;
 			for(int j = 0; j < m_dimensionNumber; j++){
 				temp_mean[j] = data[select*m_dimensionNumber+j];
 			}
-		memcpy(m_means[i], temp_mean, sizeof(double) * m_dimensionNumber);
+		memcpy(m_means[i], temp_mean, sizeof(float) * m_dimensionNumber);
 	}
 	delete[] temp_mean;
 	// finish init
 	bool loop = true;
-	double unchanged = 0;
-	double* v_data = new double[m_dimensionNumber];
+	float unchanged = 0;
+	float* v_data = new float[m_dimensionNumber];
 	int v_label = -1;
-	double cost, lastCost;
-	double** new_means = new double*[m_clusterNumber];
+	float cost, lastCost;
+	float** new_means = new float*[m_clusterNumber];
 	for (int i = 0; i < m_clusterNumber; i++)
 	{
-		new_means[i] = new double[m_dimensionNumber];
+		new_means[i] = new float[m_dimensionNumber];
 	}
 	int* new_means_counts_each_cluster = new int[m_clusterNumber];
 	while (loop){
 		cost = 0.0;
 		for (int i = 0; i < m_clusterNumber; i++)
 		{
-			memset(new_means[i], 0, sizeof(double) * m_dimensionNumber);
+			memset(new_means[i], 0, sizeof(float) * m_dimensionNumber);
 		}
 		memset(new_means_counts_each_cluster, 0, sizeof(int) * m_clusterNumber);
 	
@@ -98,7 +98,7 @@ void KMeans::Cluster(double *data, int size, int *Label)
 				for (int j = 0; j < m_dimensionNumber; j++){
 					new_means[i][j] /= new_means_counts_each_cluster[i];
 				}
-				memcpy(m_means[i], new_means[i], sizeof(double) * m_dimensionNumber);	
+				memcpy(m_means[i], new_means[i], sizeof(float) * m_dimensionNumber);	
 			}
 		}
 		//terminal conditcon
